Moved MessageCommandsModule to Quotebot.Interactions.MessageCommand namespace. Moved GetUserQuoteCount method to Quotebot.Interactions.UserCommand.UserCommandsModule
