+++ src/Quotebot/Interactions/MessageCommand/MessageCommandsModule.cs
@@ -1,9 +1,10 @@
 ﻿using Discord.Interactions;
-using System.Text;
 using Quotebot.Domain.Validators;
+using System.Text;
 
-namespace Quotebot.Interactions;
+namespace Quotebot.Interactions.MessageCommand;
 
+// ReSharper disable once UnusedType.Global
 public class MessageCommandsModule : InteractionModuleBase<SocketInteractionContext>
 {
     private readonly IDataService _dataService;
@@ -29,9 +30,9 @@ public class MessageCommandsModule : InteractionModuleBase<SocketInteractionCont
 
         if (!isValid)
         {
-           return;
+            return;
         }
-        
+
         Quoted quote = new(completeMessage)
         {
             Author = await Context.GetGuildUserName(completeMessage.Author)
@@ -50,22 +51,5 @@ public class MessageCommandsModule : InteractionModuleBase<SocketInteractionCont
             .AppendLine($"> *{quote.Author.Nickname ?? quote.Author.Username} : {completeMessage.Content}*");
 
         await FollowupAsync($"{response}");
-    }
-
-    [UserCommand("Count of Quotes")]
-    public async Task GetUserQuoteCount(IUser user)
-    {
-        await DeferAsync();
-        var guildUser = await Context.GetGuildUserName(user);
-        var countedQuotes = await _dataService.QuotesCountByUser(guildUser);
-
-        var respsonse = countedQuotes switch
-        {
-            0 => $"{guildUser.Nickname ?? guildUser.Username} has never been quoted.",
-            1 => $"{guildUser.Nickname ?? guildUser.Username} has been quoted once.",
-            _ => $"{guildUser.Nickname ?? guildUser.Username} has been quoted {countedQuotes} times."
-        };
-
-        await FollowupAsync(respsonse);
     }
 }
